% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{choice_rule}
\alias{choice_rule}
\title{Exponential choice rule}
\usage{
choice_rule(utA, utB, phi)
}
\arguments{
\item{utA}{numeric specifying the utility of option A}

\item{utB}{numeric specifying the utility of option B}

\item{phi}{numeric specifying the choice sensitivity}
}
\value{
a choice probability
}
\description{
\code{choice_rule} calculates the probability of choosing A using an ex-
  ponential choice rule.
}

