% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpt_choice}
\alias{cpt_choice}
\title{CPT-based deterministic choices}
\usage{
cpt_choice(par, problems, type = 0L)
}
\arguments{
\item{par}{numeric vector specifying the parameters of the CPT model.}

\item{problems}{a list as generated by \link{p_arrange}.}

\item{type}{integer specifying the parameterization of the value and
weighting function. See \link{v_wrapper} and \link{w_wrapper}.}
}
\value{
a vector of choices
}
\description{
\code{cpt_choice} produces deterministic choices for a set of choice
  problems and parameters.
}
