% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpt_lik}
\alias{cpt_lik}
\title{CPT-based likelihood}
\usage{
cpt_lik(par, problems, choices, type = 0L, limit = 1e-04)
}
\arguments{
\item{par}{numeric vector specifying the parameters of the CPT model.}

\item{problems}{list as generated by \link{p_arrange}.}

\item{choices}{numeric vector of 0s and 1s indicating the choices where
0 = A and 0 = B.}

\item{type}{integer specifying the parameterization of the value and
weighting function. See \link{v_wrapper} and \link{w_wrapper}.}
}
\value{
a vector of choice probabilities
}
\description{
\code{cpt_prob} computes for a set of choice problems and a set of
  parameters the combined (negative) log-likelihood of a set of choices.
}

