% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpt_prob}
\alias{cpt_prob}
\title{CPT-based choice probabilities}
\usage{
cpt_prob(par, problems, type = 0L)
}
\arguments{
\item{par}{numeric vector specifying the parameters of the CPT model.}

\item{problems}{a list as generated by \link{p_arrange}.}

\item{type}{integer specifying the parameterization of the value and
weighting function. See \link{v_wrapper} and \link{w_wrapper}.}
}
\value{
a vector of choice probabilities
}
\description{
\code{cpt_prob} computes for a set of choice problems and a set of
  parameters the respective probabilities of choosing option A.
}
