% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EXPO_choice}
\alias{EXPO_choice}
\title{EXPO deterministic choices}
\usage{
EXPO_choice(par, problems, type, random = FALSE)
}
\arguments{
\item{par}{numeric vector specifying the discounting factor
and the choice sensitivity (in that order).}

\item{problems}{a numeric matrix with four columns containing for each decision
the outcome and delay of the sooner smaller option, the outcome and delay
of the larger later option (in that order).}

\item{type}{integer specifying the model variant to be used. See details.}

\item{random}{logical specifiyng whether ties should lead to NA (here -1) or
random choice.}
}
\value{
A vector of choices.
}
\description{
Generate deterministic choice from various exponential discounting models.
}
\details{
Function takes parameter values and the choice problem to generate
deterministic choices.

Functions implements different variants as a function of \code{type}:

00 implements \eqn{k^delay} variant with exponential choice rule

01 implements \eqn{k^delay} variant with homothetic choice rule

10 implements \eqn{e^{k*delay}} variant with exponential choice rule

11 implements \eqn{e^{k*delay}} variant with homothetic choice rule

Note that for deterministic choice 00 = 01 and 10 = 11, i.e., choice rules
become irrelevant.
}
