% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EXPO_lik}
\alias{EXPO_lik}
\title{EXPO likelihood}
\usage{
EXPO_lik(par, problems, choices, epsilon = 1e-04, type = 0L)
}
\arguments{
\item{par}{numeric vector specifying the discounting factor
and the choice sensitivity (in that order).}

\item{problems}{a numeric matrix with four columns containing for each decision
the outcome and delay of the sooner smaller option, the outcome and delay
of the larger later option (in that order).}

\item{choices}{numeric vector containing the choice choice (0 = sooner smaller,
1 = larger later).}

\item{epsilon}{numeric scalar specifying a boundary on the likelihood of a choice.
Specifically, likelihoods will be bound between epsilon and 1-epsilon.}

\item{type}{integer specifying the model variant to be used. See details.}
}
\value{
The negative log-liklihood XXX.
}
\description{
Likelihood function for various exponential discounting models.
}
\details{
Function takes parameter values, the choice problem, and the actual choices
to compute a likelihood of the choices given the parameters.

Functions implements different variants as a function of \code{type}:

00 implements \eqn{k^delay} variant with exponential choice rule

01 implements \eqn{k^delay} variant with homothetic choice rule

10 implements \eqn{e^{-k*delay}} variant with exponential choice rule

11 implements \eqn{e^{-k*delay}} variant with homothetic choice rule
}
