% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EXPO_prob}
\alias{EXPO_prob}
\title{EXPO probabilities}
\usage{
EXPO_prob(par, problems, type = 0L)
}
\arguments{
\item{par}{numeric vector specifying the discounting factor
and the choice sensitivity (in that order).}

\item{problems}{a numeric matrix with four columns containing for each decision
the outcome and delay of the sooner smaller option, the outcome and delay
of the larger later option (in that order).}

\item{type}{integer specifying the model variant to be used. See details.}
}
\value{
A vector of choice probabilities.
}
\description{
Generate predicted choice probabilities from exponential discounting model with
exponential choice rule.
}
\details{
Function takes parameter values and the choice problem to generate
probabilistic predictions.

Functions implements different variants as a function of \code{type}:

00 implements \eqn{k^delay} variant with exponential choice rule

01 implements \eqn{k^delay} variant with homothetic choice rule

10 implements \eqn{e^{k*delay}} variant with exponential choice rule

11 implements \eqn{e^{k*delay}} variant with homothetic choice rule
}
