% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expo_rule}
\alias{expo_rule}
\title{Exponential choice rule}
\usage{
expo_rule(diff, phi)
}
\arguments{
\item{diff}{difference in utility between sooner smaller
and larger later options.}

\item{phi}{sensitivity parameter.}
}
\value{
Probability to choose sooner smaller option \eqn{p_ss}.
}
\description{
Exponential choice rule is defined as
\eqn{p_ss = \frac{1}{1 + e^{\phi * (u_ll - u_ss)}}}
}
