% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{homo_rule}
\alias{homo_rule}
\title{Homothetic choice rule}
\usage{
homo_rule(u_ss, u_ll, phi)
}
\arguments{
\item{u_ss}{utility of sooner smaller option.}

\item{u_ll}{utility of larger later option.}

\item{phi}{sensitivity parameter.}
}
\value{
Probability to choose sooner smaller option \eqn{p_ss}.
}
\description{
Homothetic choice rule is defined as
\eqn{p_ss = \frac{u_ss^\phi}{u_ss^\phi + u_ll^\phi}}
}
