% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HYPER_rndchoice}
\alias{HYPER_rndchoice}
\title{HYPER random choices}
\usage{
HYPER_rndchoice(par, problems, type = 0L)
}
\arguments{
\item{par}{numeric vector specifying the discounting factor
and the choice sensitivity (in that order).}

\item{problems}{a numeric matrix with four columns containing for each decision
the outcome and delay of the sooner smaller option, the outcome and delay
of the larger later option (in that order).}

\item{type}{integer specifying the model variant to be used. See details.}
}
\value{
A vector of choices.
}
\description{
Generate random choice from various hyperbolic discounting models.
}
\details{
Functions implements different variants as a function of \code{type}:

00 implements the one parameter variant with exponential choice rule

01 implements the one parameter variant with homothetic choice rule

10 implements the two parameter variant with exponential choice rule

11 implements the two parameter variant with homothetic choice rule

The delay of the one parameter variant is modeled as
\deqn{1.0 / (1.0 + k * delay)}

and that of the two parameter variant is modeled as
\eqn{(1.0 / (1.0 + k * delay))^s}
}
