% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ITCH_pgen}
\alias{ITCH_pgen}
\title{ITCH data}
\usage{
ITCH_pgen(problems)
}
\arguments{
\item{problems}{a numeric matrix with four columns containing for each decision
the outcome and delay of the sooner smaller option, the outcome and delay
of the larger later option (in that order).}
}
\value{
Four column matrix containing the difference in outcomes, the differences
in outcomes normalized by the mean of outcomes, the difference in delays, and the
difference in delays normalized by the mean of delays.
}
\description{
Generates matrix containing the four variables needed to compute
the ITCH.
}
