% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ITCH_prob}
\alias{ITCH_prob}
\title{ITCH probabilities}
\usage{
ITCH_prob(par, problems, type = 0L)
}
\arguments{
\item{par}{numeric vector specifying the discounting factor
and the choice sensitivity (in that order).}

\item{problems}{a numeric matrix with four columns containing for each decision
the outcome and delay of the sooner smaller option, the outcome and delay
of the larger later option (in that order).}

\item{type}{integer specifying the model variant to be used. See details.}

\item{choices}{numeric vector containing the choice choice (0 = sooner smaller,
1 = larger later).}
}
\value{
A vector of choices.
}
\description{
Generate predicted choice probabilities from variants of the intertemporal
choice heuristic (ITCH).
}
\details{
Function takes parameter values and the choice problem to generate
random choices.

Functions implements different model variants as a function of \code{type}:

00 implements the four parameter variant with exponential choice rule

10 implements the five parameter variant with exponential choice rule

The four and five parameter variants are defined as
\deqn{\beta_0 * (o_ss - o_ll) + \beta_1 * \frac{o_ss - o_ll}{o*} +
      \beta_2 * (d_ss - d_ll) + \beta_3 * \frac{d_ss - d_ll}{d*}}
and

\deqn{\beta_5 +
      \beta_0 * (o_ss - o_ll) + \beta_1 * \frac{o_ss - o_ll}{o*} +
      \beta_2 * (d_ss - d_ll) + \beta_3 * \frac{d_ss - d_ll}{d*}}

With o_ss, o_ll, d_ss, d_ll being the outcomes and delays of the
sooner smaller and larger later options, respectively, and o* and d*
are the arithmetic means of both outcomes and delays, respectively.

Note that the ITCH cannot be used with a homothetic choice rule.
}
