% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_arrange}
\alias{p_arrange}
\title{Rearrange problems}
\usage{
p_arrange(problems, nA = 0L)
}
\arguments{
\item{problems}{np x (2*nA + 2*nB) matrix containing the np problems each
having nA and nB outcomes.}

\item{nA}{number of outcomes in option A}
}
\value{
a list of two matrices containing the outcomes and probabilities of
  the A and B options rearranged for CPT and TAX.
}
\description{
\code{p_arrange} rearranges the matrix of problem to match the cumulative and
  decumulative specifications of CPT and TAX
}
