% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pevbyn_nm}
\alias{pevbyn_nm}
\title{Probability of choosing EV}
\usage{
pevbyn_nm(p, n)
}
\arguments{
\item{p}{numeric vector containing the outcomes of A, the respective probabilities
of the outcomes of A, the outcomes of B, and the respective probabilities of the
outcomes of B.}

\item{n}{integer specifying the sample size in terms of the number of samples taken
from \emph{each} of the options.}
}
\value{
a numeric value specifying the probability of choosing the higher EV option.
}
\description{
Function calculates the probability of choosing EV for a given a
binary problem (between options with maximally 2 outcomes) and sample size based
on a strategy that chooses the options with the higher experienced mean.
}
