% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pgen_dd}
\alias{pgen_dd}
\title{Generate delay discounting problems}
\usage{
pgen_dd(n_problem, o_lower = 0, o_upper = 100, d_lower = 0,
  d_upper = 100)
}
\arguments{
\item{n_problem}{integer specifying the number of problems.}

\item{o_lower}{integer specifying the smalles possible outcome.}

\item{o_upper}{integer specifying the largest possible outcome.}

\item{d_lower}{integer specifying the smalles possible delay.}

\item{d_upper}{integer specifying the smalles possible delay.}
}
\value{
Four column matrix containing for each problem
  the outcome and delay of the sooner smaller option and the outcome and delay
  of the larger later option (in that order).
}
\description{
Generate random delay discounting problems.
}
\details{
Algorithm draws two outcomes and two delays and then
maps the smaller outcome to the smaller delay and the larger
outcome to the larger delay.

Values are drawn from uniform.
}
