% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pgen_rnd}
\alias{pgen_rnd}
\title{Generate random decision environments}
\usage{
pgen_rnd(ns, nA = 2L, nB = 2L, lower = -100, upper = 100,
  ecological = TRUE, stdom1_test = TRUE, stdom2_test = TRUE,
  d_crit = 0.5)
}
\arguments{
\item{ns}{integer vector of length three specifying the number of choice
problems in the domain of Gain, Loss, and Mixed (in that order).}

\item{nA}{integer specifying the number of outcomes in option A}

\item{nB}{integer specifying the number of outcomes in option B}

\item{lower}{integer specifying the lower limit of the outcome range}

\item{upper}{integer specifying the upper limit of the outcome range}

\item{ecological}{bool specifies whether the outomes should be re-
arranged such that high outcomes are coupled with low probabilities}

\item{stdom1_test}{bool specifies wehther the function controls for 1st oder
stochastic dominance. If true, only non-dominant problems will be
returned.}

\item{d_crit}{double specifiying the maximum effect size between the options.
I.e., a d_crit value of .5 implies that options differences in
expected values is at most .5 times the pooled standard deviation.}

\item{stdom1_test}{bool specifies wehther the function controls for 2nd oder
stochastic dominance. If true, only non-dominant problems will be
returned.}
}
\value{
A matrix of choice problems
}
\description{
\code{pgen_rnd} generates random environments of twp-option choice
  problems without 1st and 2nd order stochastic dominance and an
  ecologically
}

