% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampl_complete}
\alias{sampl_complete}
\title{Sampling algorithm using a stop when complete stopping rule}
\usage{
sampl_complete(prob, gamma_compl, gamma_incompl)
}
\arguments{
\item{prob}{a problem table as produced by \link{p_arrange}.}

\item{gamma_compl}{a numeric controlling the likelihood with which search
stops in the case of equally many experienced outcomes. Large numbers
result in early search termination.}

\item{gamma_incompl}{a numeric controlling the likelihood with which
search stops in the case of equally many experienced outcomes. Large
numbers result in early search termination.}

\item{ns}{numceric vector whose length is either one or the number of
problems in the set.}
}
\value{
a list of witch each element containing the samples for option
  A, the samples for option B, and the choice.
}
\description{
\code{sample_complete} implements a sampling process that terminates more
  likly when (at least) an expected number of outcomes has been observed
  across both options. The expected number of outcomes is drawn from the
  distribution of number of outcomes across the entire problem set.
}

