% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampl_easy}
\alias{sampl_easy}
\title{Sampling algorithm using stop when easy stopping rule}
\usage{
sampl_easy(prob, phi, par, type)
}
\arguments{
\item{prob}{a problem table as produced by \link{p_arrange}.}

\item{phi}{a numeric specifying the sensitivity to the difference in
option utilities. Large numbers result in early search termination.}

\item{par}{a numeric vector to be passed on to \link{utility}.}

\item{type}{an integer to be passed on to \link{utility}.}

\item{ns}{numceric vector whose length is either one or the number of
problems in the set.}
}
\value{
a list of witch each element containing the samples for option
  A, the samples for option B, and the choice.
}
\description{
\code{smpl_easy} implements a sampling process that terminates sampling
  increasingly likely as the distance between the options becomes large
}
