% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampl_even}
\alias{sampl_even}
\title{Sampling algorithm using a stop when even stopping rule}
\usage{
sampl_even(prob, gamma_eq, gamma_uneq)
}
\arguments{
\item{prob}{a problem table as produced by \link{p_arrange}.}

\item{gamma_eq}{a numeric controlling the likelihood with which search
stops in the case of equally many experienced outcomes. Large number
result in early search termination}

\item{gamma_uneq}{a numeric controlling the likelihood with which search
stops in the case of equally many experienced outcomes. Large numbers
result in early search termination.}
}
\value{
a list of witch each element containing the samples for option
  A, the samples for option B, and the choice.
}
\description{
\code{sample_even} implements a sampling process that terminates more
  likly when equally many outcomes in both options have been experienced.
}

