% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampl_n}
\alias{sampl_n}
\title{Sample a fixed number of outcomes per option}
\usage{
sampl_n(prob, ns)
}
\arguments{
\item{prob}{a problem table as produced by \link{p_arrange}.}

\item{ns}{numceric vector whose length is either one or the number of
problems in the set.}
}
\value{
a list of samples.
}
\description{
\code{sampl_n} draws a fixed number of samples for each option in the
  set of problems.
}
