% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampl_sure}
\alias{sampl_sure}
\title{Sample using a search until rare event}
\usage{
sampl_sure(prob, gamma_eq, gamma_uneq)
}
\arguments{
\item{prob}{a problem table as produced by \link{p_arrange}.}

\item{gamma_eq}{a numeric controlling the likelihood with which search
stops in the case of equally many experienced outcomes.}

\item{gamma_uneq}{a numeric controlling the likelihood with which search
stops in the case of equally many experienced outcomes.}

\item{ns}{numceric vector whose length is either one or the number of
problems in the set.}
}
\value{
a list of witch each element containing the samples for option
  A, the samples for option B, and the choice.
}
\description{
\code{sample_swe} implements a sampling process that terminates more
  likly when equally many outcomes in both options have been experienced.
}

