% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampl_swe}
\alias{sampl_swe}
\title{Sample using a stop when easy rule}
\usage{
sampl_swe(prob, phi, par, type)
}
\arguments{
\item{prob}{a problem table as produced by \link{p_arrange}.}

\item{phi}{a numeric specifying the sensitivity to the difference in
option utilities.}

\item{par}{a numeric vector to be passed on to \link{utility}.}

\item{type}{an integer to be passed on to \link{utility}.}

\item{ns}{numceric vector whose length is either one or the number of
problems in the set.}
}
\value{
a list of witch each element containing the samples for option
  A, the samples for option B, and the choice.
}
\description{
\code{sample_swe} implements a sampling process that terminates
  increasingly likely as the distance when the difference between the
  options is large
}

