% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stdom1}
\alias{stdom1}
\title{Test for 1st order stochastic dominance}
\usage{
stdom1(oA, pA, oB, pB)
}
\arguments{
\item{A}{numeric vector containing the outcomes and probabilities
(in that order) of option A}

\item{B}{numeric vector containing the outcomes and probabilities
(in that order) of option B}
}
\value{
A boolean indicating the status of 1st order stochastic dominance
}
\description{
\code{stdom1} tests if one of two problems is stochastically dominant
  according to 1st order stochastic dominance.
}

