% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stdom2}
\alias{stdom2}
\title{Test for 2nd order stochastic dominance}
\usage{
stdom2(oA, pA, oB, pB, d_crit = 0)
}
\arguments{
\item{A}{numeric vector containing the outcomes and probabilities
(in that order) of option A}

\item{B}{numeric vector containing the outcomes and probabilities
(in that order) of option B}
}
\value{
A boolean indicating the status of 2nd order stochastic dominance
}
\description{
\code{stdom2} tests if one of two problems is stochastically dominant
  according to 2nd order stochastic dominance.
}

