% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{t}
\alias{t}
\title{Weighting function of TAX}
\usage{
t(p, gamma)
}
\arguments{
\item{p}{numeric specifying the to be transformed probability.}

\item{gamma}{numeric specifying the exponential factor
(sensitivity) of the probability weighting function.}
}
\value{
a decision weight
}
\description{
\code{t} is the weighting function proposed by Tverksy & Kahneman
  (1992). It transforms probabilities into decision weights.
}
\references{
Birnbaum, M. H. (1999). Testing critical properties of
  decision making on the Internet. Psychological Science, 10(5), 399-407.
}

