% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{t_tax}
\alias{t_tax}
\title{Weighting function of TAX}
\usage{
t_tax(p, gamma)
}
\arguments{
\item{p}{numeric specifying the to be transformed probability.}

\item{gamma}{numeric specifying the exponential factor
(sensitivity) of the probability weighting function.}
}
\value{
a decision weight
}
\description{
\code{t_tax} is the weighting function of TAX.
}
\references{
Birnbaum, M. H. (1999). Testing critical properties of
  decision making on the Internet. Psychological Science, 10(5), 399-407.
}

