% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tax_choice}
\alias{tax_choice}
\title{TAX-based deterministic choices}
\usage{
tax_choice(par, problems, type = 1L)
}
\arguments{
\item{par}{numeric vector specifying the parameters of the CPT model.}

\item{problems}{a list as generated by \link{p_arrange}.}

\item{type}{integer specifying the parameterization of the tax model
function. See \link{utility_tax}.}
}
\value{
a vector of choices
}
\description{
\code{tax_choice} produces deterministic choices for a set of choice
  problems and parameters.
}
