% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tax_lik}
\alias{tax_lik}
\title{TAX-based likelihood}
\usage{
tax_lik(par, problems, choices, type = 1L, limit = 1e-04)
}
\arguments{
\item{par}{numeric vector specifying the parameters of the TAX model.}

\item{problems}{list as generated by \link{p_arrange}.}

\item{choices}{numeric vector of 0s and 1s indicating the choices where
0 = A and 0 = B.}

\item{type}{integer specifying the parameterization of the tax model. See
\link{utility_tax}.}
}
\value{
a vector of choice probabilities
}
\description{
\code{tax_prob} computes for a set of choice problems and a set of
  parameters the combined (negative) log-likelihood of a set of choices.
}

