% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tax_prob}
\alias{tax_prob}
\title{TAX-based choice probabilities}
\usage{
tax_prob(par, problems, type = 1L)
}
\arguments{
\item{par}{numeric vector specifying the parameters of the TAX model.}

\item{problems}{a list as generated by \link{p_arrange}.}

\item{type}{integer specifying the parameterization of tax model. See
\link{utility_tax}.}
}
\value{
a vector of choice probabilities
}
\description{
\code{tax_prob} computes for a set of choice problems and a set of
  parameters the respective probabilities of choosing option A.
}
