% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{u}
\alias{u}
\title{Value function of TAX}
\usage{
u(o, alpha)
}
\arguments{
\item{o}{numeric specifying the magnitude of the to be transformed
outcome}

\item{alpha}{numeric specifying the exponent of the utility function for
the gain domain.}
}
\value{
a utility
}
\description{
\code{u}, the TAX value function, transforms the magnitute of
  an outcome into utility space.
}

