% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{utility_tax}
\alias{utility_tax}
\title{Utility function of TAX}
\usage{
utility_tax(opt, par, type = 1L)
}
\arguments{
\item{opt}{numeric vector specifying the outcomes and probabilities of an
option. The function expects a length of (number of outcomes * 2 + 1),
that is ordered according to max_loss, min_loss, max_gain, min_gain, and
whose last entry indicates the number of loss outcomes.}

\item{par}{numeric vector specifying the parameters of the CPT model. See
\link{u}, \link{t_tax}, and \link{w}.}

\item{type}{integer specifying the parameterization of TAX.
\code{type = 0} is the one parameter TAX with just a \code{gamma}
parameter. \code{type = 1} is the two parameter TAX with a \code{gamma}
and a \code{alpha} parameter. \code{type = 2} is the two parameter TAX
with a \code{gamma} and a \code{delta} parameter. \code{type = 3} is the
three parameter TAX with a \code{gamma}, a \code{alpha}, and a
\code{delta} parameter.}
}
\value{
a utility
}
\description{
\code{t} is the weighting function proposed by Tverksy & Kahneman
  (1992). It transforms probabilities into decision weights.
}
\references{
Birnbaum, M. H. (1999). Testing critical properties of
  decision making on the Internet. Psychological Science, 10(5), 399-407.
}

