% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{v}
\alias{v}
\title{Standard value function}
\usage{
v(o, alpha, beta, lambda)
}
\arguments{
\item{o}{numeric specifying the magnitude of the to be transformed
outcome}

\item{alpha}{numeric specifying the exponent of the utility function for
the gain domain.}

\item{beta}{numeric specifying the exponent of the utility function for
the loss domain.}

\item{lambda}{numeric specifying the slope of the utility function for
the loss domain relative to the gain domain.}
}
\value{
a utility
}
\description{
\code{v}, being the standard value function, transforms the magnitute of
  an outcome into utility space
}

