% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{w}
\alias{w}
\title{Transfer function of TAX}
\usage{
w(pi, pk, delta, gamma, n)
}
\arguments{
\item{pi}{numeric specifying the to be transformed probability at i.}

\item{pk}{numeric specifying the to be transformed probability at k.}

\item{delta}{numeric specifying the transfer factor.}

\item{gamma}{numeric specifying the exponential factor (sensitivity) of
the probability weighting function. See \link{t_tax}.}
}
\value{
a transfer weight
}
\description{
\code{w} is the transfer function of TAX.
}
\references{
Birnbaum, M. H. (1999). Testing critical properties of
  decision making on the Internet. Psychological Science, 10(5), 399-407.
}
