% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{w_ge}
\alias{w_ge}
\title{Weighting function according to Goldstein & Einhorn (1987)}
\usage{
w_ge(p, o, delta_l, delta_g, gamma_l, gamma_g)
}
\arguments{
\item{p}{numeric specifying the to be transformed probability.}

\item{delta_l, delta_g}{numeric specifying the multiplicative factor
(elevation) of the weighting function for the loss (l) and gain
(g) domain.}

\item{gamma_l, gamma_g}{numeric specifying the exponential factor
(sensitivity) of the weighting function for the loss (l) and gain
(g) domain.}
}
\value{
a decision weight
}
\description{
\code{w_ge} is the weighting function proposed by Goldstein & Einhorn
  (1987). It transforms probabilities into decision weights.
}
\references{
Goldstein, W. M. and Einhorn, H. J. (1987). Expression Theory
  and the Preference Reversal Phenomena. Psychological Review 94, 236—254.
}

