% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{w_tk}
\alias{w_tk}
\title{Weighting function according to Tversky & Kahneman (1992)}
\usage{
w_tk(p, o, gamma_l, gamma_g)
}
\arguments{
\item{p}{numeric specifying the to be transformed probability.}

\item{gamma_l, gamma_g}{numeric specifying the exponential factor
(sensitivity) of the weighting function for the loss (l) and gain
(g) domain.}
}
\value{
a decision weight
}
\description{
\code{w_tk} is the weighting function proposed by Tverksy & Kahneman
  (1992). It transforms probabilities into decision weights.
}
\references{
Tversky, A., & Kahneman, D. (1992). Advances in prospect
  theory: Cumulative representation of uncertainty. Journal of Risk and
  uncertainty, 5(4), 297-323.
}

