% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{w_wrapper}
\alias{w_wrapper}
\title{Wrapper for weighting function}
\usage{
w_wrapper(p, o, par, type)
}
\arguments{
\item{p}{numeric specifying the to be transformed probability.}

\item{o}{numeric specifying the outcome associated with the probability.}

\item{type}{integer specifying the parameterization of the weighting
function. See Details.}
}
\value{
a decision weight
}
\description{
\code{w_wrapper} is a wrapper for different parameterizations of the
  weighting function.
}

