% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{v_wrapper}
\alias{v_wrapper}
\title{Wrapper for value function}
\usage{
v_wrapper(o, par, type)
}
\arguments{
\item{o}{numeric specifying the to be transformed outcome.}

\item{type}{integer specifying the parameterization of the value function.
See Details.}
}
\value{
a utility
}
\description{
\code{v_wrapper} is a wrapper for different parameterizations of the value
  function.
}
\details{
The type argument controls the type of value function implemented.

type = xx1 implements only \code{alpha}

type = xx2 implements \code{alpha} and \code{beta}

type = xx3 implements \code{alpha}, \code{beta}, and \code{lambda}
}

