% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{w_wrapper}
\alias{w_wrapper}
\title{Wrapper for weighting function}
\usage{
w_wrapper(p, o, par, type)
}
\arguments{
\item{p}{numeric specifying the to be transformed probability.}

\item{o}{numeric specifying the outcome associated with the probability.}

\item{type}{integer specifying the parameterization of the weighting
function. See Details.}
}
\value{
a decision weight
}
\description{
\code{w_wrapper} is a wrapper for different parameterizations of the
  weighting function.
}
\details{
The \code{type} argument controls the type of weighting function implemented.

00x implements the weighting function of Tversky & Kahneman
(1992) with one \code{gamma}.

01x implements the weighting function of Tversky & Kahneman
(1992) with one \code{gamma} for losses and one \code{gamma} for gains.

10x implements the weighting function of Goldstein & Einhorn
(1987) with one \code{delta} and one \code{gamma}.

11x implements the weighting function of Goldstein & Einhorn
(1987) with one \code{delta}, one \code{gamma} for losses, and one
\code{gamma} for gains.

12x implements the weighting function of Goldstein & Einhorn
(1987) with one \code{delta} for losses, one \code{delta} for gains,
and one \code{gamma}.

13x implements the weighting function of Goldstein & Einhorn
(1987) with one \code{delta} for losses, one \code{delta} for gains,
and one \code{gamma} for losses and one \code{gamma} for gains.

20x implements the "one"-parameter weighting function of Prelec
(1998) with one \code{gamma}.

21x implements the "one"-parameter weighting function of Prelec
(1998) with one \code{gamma} for losses and one \code{gamma} for gains.

22x implements the "two"-parameter weighting function of Prelec
(1998) with one \code{delta} and one \code{gamma}.

23x implements the "two"-parameter weighting function of Prelec
(1998) with one \code{delta}, one \code{gamma} losses, and one
\code{gamma} for gains.

24x implements the "two"-parameter weighting function of Prelec
(1998) with one \code{delta} for losses, one \code{delta} gains, and
one \code{gamma}.

25x implements the "two"-parameter weighting function of Prelec
(1998) with one \code{delta} for losses, one \code{delta} gains, one
\code{gamma} for losses and one \code{gamma} for gains.
}
